- added unsupported Horizon RAMDISK emulation [ramdisk] disk=D:\classic99\ramdisk.bin size=8388608
The presence of "disk" enables it, size must be reasonable as well. Tested up to 16MB, I think that's the max.
The RAMdisk will override any Classic99 disk. All data is just dumped into that file. The first 8k of the file
is the DSR, the rest is the raw data written to RAM.

Note CFG is pretty slow when there are megabytes of disk to size. Expect a few seconds.

If you change the size, you should delete the file as well. Only one RAMdisk at >1000 is supported, and you can
not use a CF7 at the same time. RAMBO is not supported.

- fix ignoring illegal VDP register writes (they should be masked)
- fix F18A 5S flag when more than 4 sprites per line are enabled
- make CF7 functions static where possible

a